uvicore/auth/database/tables/permissions.py
import uvicore
import sqlalchemy as sa
from uvicore.database import Table
from uvicore.support.dumper import dump, dd


@uvicore.table()
class Permissions(Table):

    # Actual database table name
    # Plural table names and singluar model names are encouraged
    # Do not add a package prefix, leave that to the connection config
    name = 'permissions'

    # Connection for this database from your config file
    connection = 'auth'

    # SQLAlchemy Table definition as a list (exclude name and metadata)
    # This will be converted into an actual SQLAlchemy Table() instance
    # See https://docs.sqlalchemy.org/en/13/core/schema.html
    schema = [
        sa.Column('id', sa.Integer, primary_key=True),
        sa.Column('entity', sa.String(length=50), index=True),
        sa.Column("name", sa.String(length=100))
    ]

    # Optional SQLAlchemy Table() instance kwargs
    schema_kwargs = {}
