pyproject.toml
[tool.poetry]
name = "uvicore"
version = "0.1.6"
license = "MIT"
authors = ["Matthew Reschke <mail@mreschke.com>"]
description = "The Async Python Framework for Artisans. An Elegant Fullstack Python Web, API and CLI Framework"
readme = "README.md"
homepage = "https://github.com/uvicore/framework"
documentation = "https://github.com/uvicore/framework"
repository = "https://github.com/uvicore/framework"
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
include = ["LICENSE"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
fastapi = "^0.63.0"
uvicorn = "^0.13.3"
gunicorn = "^20.0.4"
aiofiles = "^0.6.0"
databases = "^0.4.1"
alembic = "^1.5.2"
asyncclick = "^7.1.2"
colored = "^1.4.2"
prettyprinter = "^0.18.0"
environs = "^9.3.0"
Jinja2 = "^2.11.2"
ipython = "^7.19.0"
requests = "^2.25.1"
itsdangerous = "^1.1.0"
uvloop = "^0.14.0"
httptools = "^0.1.1"
python-multipart = "^0.0.5"
argon2-cffi = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
pytest-cov = "^2.11.1"
pytest-asyncio = "^0.14.0"
black = "^20.8b1"
databases = {extras = ["mysql", "sqlite"], version = "^0.4.1"}
mkdocs = "^1.1.2"
mkdocs-material = "^6.2.6"
jupyterlab = "^3.0.5"
aioredis = "^1.3.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
